Add requirements.txt, change item scores from int to double, distribute typeItem score equally among its items

+++ java-exam-evaluator-annotations/src/main/java/hu/sed/evaluator/annotation/semantic/CustomTest.java
@@ -39,5 +39,5 @@ public @interface CustomTest {
      *
      * @return maximum score
      */
-    int score() default 1;
+    double score() default 1;
 }
