+++ java-exam-evaluator-annotations/src/main/java/hu/sed/evaluator/annotation/syntax/FieldCheck.java
@@ -47,5 +47,5 @@ public @interface FieldCheck {
      *
      * @return maximum score
      */
-    int score() default 1;
+    double score() default 1;
 }
