+++ java-exam-evaluator-annotations/src/main/java/hu/sed/evaluator/annotation/syntax/MethodCheck.java
@@ -62,5 +62,5 @@ public @interface MethodCheck {
      *
      * @return maximum score
      */
-    int score() default 1;
+    double score() default 1;
 }
